PATHFINDING AND TURNS IMPLEMENTED
IT WORKS! However there's still a lot of softlocks and the logic is unstable. Monsters get stuck on each other and don't avoid each other. They can also get misaligned with the tiles when attempting to make what's supposed to be an invalid movement. Killing monsters has inconsistent behavior.

Being able to pass turns, eat monsters, and a functioning movement resource are the last major things for the player side of the gameplay loop to be minimally functioning. Monsters still need a means of attack.