Basic monster attack implemented (hardcoded)
Not only has a one space large melee attack been implemented, monsters can correctly decide to attack the player and deal damage when in range. The current solution is not scalable for having more attacks available with different ranges and must be replaced later.

The game over condition of the gameplay loop has been properly established. The turn system still needs extra functionality to wait for attacks to complete, and the player collision could be polished up. Monsters still stack up on each other and get stuck with their pathfinding.

The last requirement for the core gameplay loop is a calorie-free attack from the player.