Beginning level scripting
The existing lock picture was too big so I resized it to not conflict with tile maps. The tileset resources need to be a lot cleaner when real levels are created. The new battle_test level will help with debugging general level progression, such as unlocking doors, solving puzzles, and encountering enemies.