Event trigger and puzzle system work
More functionality to the event trigger box. Puzzle signals can be connected and emitted to puzzle objects similar to an electric circuit. This includes a basic puzzle component that listens for puzzle signals and calls a 'solve' function when the conditions are met. The included default behavior doesn't work for disabling TileMapLayer collision, which is what the locked door tiles are made from. A unique script for progression locking doors is necessary.